increase char limit, implemented delete all query and command

+++ internal/srs/delete.go
@@ -6,6 +6,14 @@ import (
 )
 
 func (s *State) DeleteEntry(id int64) error {
+	fmt.Printf("Delete entry #%d? [y/N] ", id)
+	var confirm string
+	fmt.Scanln(&confirm)
+	if confirm != "y" && confirm != "Y" {
+		fmt.Println("Delete aborted.")
+		return nil
+	}
+
 	result, err := s.DB.DeleteEntry(context.Background(), id)
 	if err != nil {
 		return fmt.Errorf("couldn't delete entry: %w", err)
@@ -22,4 +30,22 @@ func (s *State) DeleteEntry(id int64) error {
 
 	fmt.Printf("[%d] Deleted Successfully\n", id)
 	return nil
+}
+
+func (s *State) RemoveAllEntry() error {
+	fmt.Printf("⚠️ WARNING, all entries will be permanently deleted. Are you sure? [y/N] ", )
+	var confirm string
+	fmt.Scanln(&confirm)
+	if confirm != "y" && confirm != "Y" {
+		fmt.Println("Delete aborted.")
+		return nil
+	}
+
+    rowsDeleted, err := s.DB.DeleteAllEntries(context.Background())
+    if err != nil {
+        return fmt.Errorf("couldn't delete entries: %w", err)
+    }
+
+    fmt.Printf("%d entries deleted\n", rowsDeleted)
+    return nil
 }
